requirements.txt
# RAG Rules Bot - Production dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.0.1
pinecone>=3.0.0
python-dotenv>=1.0.0
pypdf>=3.0.0
